Remove arguments at top of Dockerfiles
Removed arguments at the top of the Dockerfiles before the first FROM
build layer. These arguments were previous defined above the FROM with
a default value which would allow it to be overriden based on passed in
build arguments. However, with how we are distributing the artifact
Dockerfiles, this was unneccessary.

Signed-off-by: Anderson Chauphan <achauph@sandia.gov>