Upgrade CUDA12 container to GCC12
Need a newer compiler due to a compound issue between `nvcc_wrappper`,
`nvcc`, and `cmake`.

`cmake` tries to pass `-std=gnu++2a` to `GCC@10.4.0`, but it's really
going to `nvcc_wrapper`, which changes it to `-std=c++2a` and passes it
to `nvcc`, which does not accept that option.  Upgrading to `GCC@12`
causes `cmake` to pass `-std=gnu++20` instead (full C++20 support) which
translates into `-std=c++20` which IS accepted.

Signed-off-by: Samuel E. Browne <sebrown@sandia.gov>